Reorganized packages/classes
