+++ src/test/java/ui/steps/Hooks/Hooks.java
@@ -1,4 +1,4 @@
-package ui.steps;
+package ui.steps.Hooks;
 
 import io.cucumber.java.After;
 import io.cucumber.java.Scenario;
